ejercicio 4.21: los blogs solo pueden ser eliminados por el creador gracias al token unico

+++ part4/bloglist/controllers/blogs.js
@@ -57,8 +57,30 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    if (!request.token) {
+      return response.status(401).json({ error: 'token missing' })
+    }
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).end()
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(403).json({ error: 'forbidden' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
+
     response.status(204).end()
+
   } catch (error) {
     next(error)
   }
